requirements.txt
blis==0.4.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cymem==2.0.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.17.2
plac==0.9.6
preshed==3.0.2
pyparsing==2.4.2
python-decouple==3.1
rdflib==4.2.2
requests==2.22.0
six==1.12.0
spacy==2.2.1
SQLAlchemy==1.3.9
srsly==0.1.0
thinc==7.1.1
tqdm==4.36.1
urllib3==1.25.6
wasabi==0.2.2
Werkzeug==0.16.0
PyMySQL
Flask-Heroku
gunicorn
psycopg2
